+++ example/lib/src/error_handler.dart
@@ -3,6 +3,8 @@ import 'package:flutter/foundation.dart';
 import '../logger.dart';
 
 class ErrorHandler {
+  const ErrorHandler._();
+
   static void init() {
     FlutterError.onError = ErrorHandler._recordFlutterError;
     logger.log(Log.info, title: 'ErrorHandler init');
@@ -15,6 +17,4 @@ class ErrorHandler {
   static void _recordFlutterError(FlutterErrorDetails error) {
     logger.log(Log.error, error: error, stack: error.stack);
   }
-
-  const ErrorHandler._();
 }