+++ lib/src/log_formatter/formatters/queue_formatter.dart
@@ -1,19 +1,18 @@
 import '../../support/log_event.dart';
 import '../../support/log_settings.dart';
 
-/// A [QueueFormatter] formats a list of [LogPart]s.
-abstract class QueueFormatter {
+/// A [IQueueFormatter] formats a list of [LogPart]s.
+abstract class IQueueFormatter {
   List<LogPart> format(LogEvent event, List<LogPart> list);
-
-  /// Default implementation of [QueueFormatter]. Removes unnecessary [LogPart] if there is no data for them.
-  factory QueueFormatter() => DefaultQueueFormatter();
 }
 
-class DefaultQueueFormatter implements QueueFormatter {
+/// Default implementation of [QueueFormatter]. Removes unnecessary [LogPart] if there is no data for them.
+class QueueFormatter implements IQueueFormatter {
+  @override
   List<LogPart> format(LogEvent event, List<LogPart> list) {
-    List<LogPart> queue = [];
+    final queue = <LogPart>[];
 
-    for (LogPart part in list) {
+    for (final part in list) {
       switch (part) {
         case LogPart.error:
           if (event.error == null) continue;