+++ lib/src/support/log_event.dart
@@ -1,19 +1,19 @@
 import 'log_type.dart';
 
 class LogEvent {
-  final dynamic message;
-  final LogType type;
-  final String? title;
-  final dynamic error;
-  final StackTrace? stack;
-  final DateTime time;
-
   LogEvent(
     this.type, {
+    required this.time,
     this.title,
     this.error,
     this.stack,
     this.message,
-    required this.time,
   });
+
+  final dynamic message;
+  final ILogType type;
+  final String? title;
+  final dynamic error;
+  final StackTrace? stack;
+  final DateTime time;
 }