+++ lib/src/support/log_settings.dart
@@ -11,16 +11,8 @@ enum LogPart {
   divider,
 }
 
-class LogTypeSettings {
-  final LogSettings settings;
-  final Map<LogType, LogSettings> typeSettings;
-
-  const LogTypeSettings(this.settings, this.typeSettings);
-
-  LogSettings operator [](LogType key) {
-    return typeSettings[key] ?? settings;
-  }
-}
+/// A function that returns the settings for the logger.
+typedef SettingsBuilder = LogSettings Function(ILogType logType);
 
 /// Settings for the logger.
 class LogSettings {
@@ -53,6 +45,9 @@ class LogSettings {
 
   /// The number of methods to print in the error stack trace.
   final int errorStackTraceMethodCount;
+
+  /// Whether to remove <asynchronous suspension> from the stack trace.
+  final bool removeAsynchronousSuspensionFromStackTrace;
 
   /// The length of the line.
   final int lineLength;
@@ -88,6 +83,7 @@ class LogSettings {
     this.stackTraceMethodCount = 2,
     this.errorStackTraceMethodCount = 8,
     this.stackTraceBeginIndex = 2,
+    this.removeAsynchronousSuspensionFromStackTrace = true,
     this.lineLength = 120,
     this.decorateLogTypeLabel = true,
     this.selectError = true,