+++ test/log_formatter_test.dart
@@ -3,23 +3,24 @@ import 'package:test/test.dart';
 
 void main() {
   test('Test formatting of log event with default settings', () {
-    var formatter = LogFormatter(
-      LogTypeSettings(
-        LogSettings(),
-        {},
-      ),
+    final formatter = LogFormatter(
+      (_) => LogSettings(),
     );
-    var logEvent =
-        LogEvent(Log.warning, title: 'Test log event', time: DateTime.now());
-    var formattedLogEvent = formatter.format(logEvent);
+    final logEvent = LogEvent(
+      LogType.warning,
+      title: 'Test log event',
+      time: DateTime.now(),
+    );
+    final formattedLogEvent = formatter.format(logEvent);
 
-    expect(formattedLogEvent.log, equals(Log.warning));
+    expect(formattedLogEvent.log, equals(LogType.warning));
     expect(
-        formattedLogEvent.queue,
-        equals([
-          LogPart.stack,
-          LogPart.time,
-        ]));
+      formattedLogEvent.queue,
+      equals([
+        LogPart.stack,
+        LogPart.time,
+      ]),
+    );
     expect(formattedLogEvent.title, isNotNull);
     expect(formattedLogEvent.error, isNull);
     expect(formattedLogEvent.stack, isNotNull);
@@ -28,7 +29,7 @@ void main() {
   });
 
   test('Test formatting of log event with custom settings', () {
-    var settings = LogSettings(
+    final settings = LogSettings(
       logParts: [
         LogPart.time,
         LogPart.message,
@@ -37,23 +38,24 @@ void main() {
       printEmoji: false,
     );
 
-    var formatter = LogFormatter(LogTypeSettings(settings, {}));
-    var logEvent = LogEvent(
-      Log.warning,
+    final formatter = LogFormatter((_) => settings);
+    final logEvent = LogEvent(
+      LogType.warning,
       title: 'Test log event',
       message: 'Test log event message',
       time: DateTime.now(),
     );
 
-    var formattedLogEvent = formatter.format(logEvent);
+    final formattedLogEvent = formatter.format(logEvent);
 
-    expect(formattedLogEvent.log, equals(Log.warning));
+    expect(formattedLogEvent.log, equals(LogType.warning));
     expect(
-        formattedLogEvent.queue,
-        equals([
-          LogPart.time,
-          LogPart.message,
-        ]));
+      formattedLogEvent.queue,
+      equals([
+        LogPart.time,
+        LogPart.message,
+      ]),
+    );
 
     expect(formattedLogEvent.title, isNotNull);
     expect(formattedLogEvent.error, isNull);